requirements.txt
anthropic==0.96.0
httpx==0.28.1
langgraph==1.1.8
python-dotenv==1.2.2